fix types export
