+++ packages/json-rpc/ws-provider/src/node.ts
@@ -1,9 +1,8 @@
 import { WebSocket } from "ws"
-import { getInternalWsProvider, type GetWsProviderInput } from "./ws-provider"
+import { getInternalWsProvider } from "./ws-provider"
 export type { JsonRpcProvider } from "@polkadot-api/json-rpc-provider"
 
-export type * from "./ws-provider"
-export { WsEvent } from "./ws-provider"
+export * from "./types"
 
 class WS extends WebSocket {
   close() {
@@ -11,6 +10,6 @@ class WS extends WebSocket {
   }
 }
 
-export const getWsProvider: GetWsProviderInput = getInternalWsProvider(
+export const getWsProvider = getInternalWsProvider(
   WS as unknown as typeof globalThis.WebSocket,
 )
