+++ packages/json-rpc/ws-provider/tsconfig.json
@@ -3,7 +3,6 @@
   "include": ["src", "tests"],
   "compilerOptions": {
     "baseUrl": "src",
-    "stripInternal": true,
     "paths": {
       "@/*": ["*"]
     }
